Split out cstr as well
